restarted cause last tutorial was trash
